+++ packages/react-craft-ai-decision-tree/src/components/decisionTree.jsx
@@ -34,14 +34,14 @@ const DecisionTree = ({ data, height, width, scale, position, updatePositionAndZ
 
 DecisionTree.defaultProps = {
   position: [0, 0],
-  scale: 1,
+  scale: -1,
   auto: false,
   updatePositionAndZoom: null
 };
 
 DecisionTree.propTypes = {
-  scale: PropTypes.number.isRequired,
-  position: PropTypes.array.isRequired,
+  scale: PropTypes.number,
+  position: PropTypes.array,
   data: PropTypes.object.isRequired,
   height: PropTypes.number,
   width: PropTypes.number,